[CCUBE-303][RL] Organise calendar files, initial clean up
