+++ src/shared/internal-calendar/day-of-month/calendar-day-view.style.tsx
@@ -1,7 +1,7 @@
 import styled, { css } from "styled-components";
-import { Color } from "../../color";
-import { Text, TextStyleHelper } from "../../text";
-import { DayVariant } from "./internal-calendar-day";
+import { Color } from "../../../color";
+import { Text, TextStyleHelper } from "../../../text";
+import { DayVariant } from "./calendar-day-view";
 
 // =============================================================================
 // STYLE INTERFACES, transient props are denoted with $
