+++ src/time-slot-week-view/time-slot-week-days.style.tsx
@@ -1,6 +1,6 @@
 import styled, { css } from "styled-components";
 import { Color } from "../color";
-import { DayLabel } from "../shared/internal-calendar/internal-calendar-day.style";
+import { DayLabel } from "../shared/internal-calendar/day-of-month/calendar-day-view.style";
 import { TextStyleHelper } from "../text";
 
 export const DayLabelWeek = styled(DayLabel)`
